Class FcastModel

Class FcastModel

Direct Superclass

NVDModel

Direct Subclasses

BenchmarkModel,

MeanModel,

DriftModel,

NaiveModel,

SeasonalNaiveModel

Private attributes

box_cox_applied logical value indicating whether the Box-Cox transformation was applied (default F)

Public methods

considerBoxCox(fcast, fcast_bc) selects the best forecasts between those where the Box-Cox transformation is used and those, where it is not used; sets thre result by using super$setFcasted

analyseResiduals(df) analyses the residuals with the given degree of freedom value

considerBootstrap(fcast, fcast_bc) generates forecast intervals from bootstrapped residuals when super$areResidualsNotNormal() & !super$areResidualsNotRandom(); takes into account whether the Box-Cox transformation was beneficial to use or not

executeUseModel(fcast, fcast_bc) calls the forecast function with training set obtained by calling super$getMergedTrainTestSet(); applies the Box-Cox transformation when self$isBoxCoxApplied() == TRUE

isBoxCoxApplied() returns private attribute box_cox_applied

setBoxCoxApplied(logical_value) assings the logical_value to private attribute box_cox_applied