Class StructTSModel

Represents basic structural models (BSM).

Direct Superclass

FitModel

Important Dependencies

stats::StructTS

forecast::forecast

Public methods

buildModel() passes on fitting and forecasting functions to FitModel's method setFittedFcasted; ignores passing function which checks whether residuals are not independently distributed

useModel(fcast_period) calls FitModel's method executeUseModel to fit model and forecast fcast_period-step ahead future; passes on the same fitting and forecasting functions as buildmodel()