Reported Vulnerabilities's Data from National Vulnerability Database (NVD)

A data set obtained by processing NVD XML files corresponding to CVE ID entries from CVE-2011 to CVE-2016 (included).

nvd

Format

A data.table::data.table with 11 variables and 41190 rows:

cve_id

ID assigned by Common Vulnerabilities and Exposures (CVE)

cve_rejected

TRUE if vulnerability is rejected by the CVE editors

published

Date when the vulnerability got published according to the vulnerability entry's XML tag vuln:published-datetime

cwe

Common Weakness Enumeration (CWE) vulnerability type. NVD uses a subset of CWEs, meaning some vulnerabilities might have type NA.

cvss_score

Common Vulnerability Scoring System (CVSS) version 2.0 Base Score

cvss_av

Access Vector (AV) [Local (L), Adjacent Network (A), Network (N)] used to calculate Base Score

cvss_ac

Access Complexity (AC) [High (H), Medium (M), Low (L)] used to calculate Base Score

cvss_au

Authentication (Au) [Multiple (M), Single (S), None (N)] used to calculate Base Score

cvss_c

Confidentiality Impact (C) [None (N), Partial (P), Complete (C)]used to calculate Base Score

cvss_i

Integrity Impact (I) [None (N), Partial (P), Complete (C)] used to calculate Base Score

cvss_a

Availability Impact (A) [None (N), Partial (P), Complete (C)] used to calculate Base Score

Source

XML Version 2.0 from https://nvd.nist.gov/vuln/data-feeds#CVE_FEED as of 7 October 2017,

NVD's CVSS introduction: https://nvd.nist.gov/vuln-metrics/cvss,

A Complete Guide to the Common Vulnerability Scoring System Version 2.0, Base-Metrics: https://www.first.org/cvss/v2/guide#2-1-Base-Metrics.